# db-viewer-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A simple API to upload CSVs to SQLite and run queries.",
        "title": "Database Visualizer API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/db-info": {
            "get": {
                "description": "Returns all tables, their rows, and inferred relationships based on column names (e.g., user_id -\u003e users)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "handleDBInfo"
                ],
                "summary": "Get Database Schema \u0026 Data",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/query": {
            "post": {
                "description": "Executes a raw SQL query against the in-memory database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "QueryExecuter"
                ],
                "summary": "Run SQL Query",
                "parameters": [
                    {
                        "description": "SQL Query",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.QueryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/upload": {
            "post": {
                "description": "Uploads a CSV file and creates a table in SQLite",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DataFileUpload"
                ],
                "summary": "Upload CSV",
                "parameters": [
                    {
                        "type": "file",
                        "description": "CSV File",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.QueryRequest": {
            "type": "object",
            "properties": {
                "query": {
                    "type": "string",
                    "example": "SELECT * FROM users"
                }
            }
        }
    }
}